build: clear two pre-existing warnings (sign-conversion, shadowed kScales)
- PluginProcessor.cpp: dsp[(size_t) c] — dsp is a std::array (operator[] takes
  size_t); the int channel index c triggered -Wsign-conversion. c stays int for
  the JUCE channel APIs; cast only at the array index.
- PluginEditor.cpp: the scale-menu lambda re-declared a static constexpr kScales
  identical to the one in the enclosing function (-Wshadow-uncaptured-local).
  Removed it; the lambda references the outer static constexpr directly.

Both behaviour-identical. Full build now 0 warnings / 0 errors.

+++ src/PluginProcessor.cpp
@@ -181,7 +181,7 @@ void TommyAudioProcessor::processBlock (juce::AudioBuffer<float>& buffer, juce::
             inPeak[c] = juce::jmax (inPeak[c], std::abs (wet));
         }
 
-        dsp[c].processBlock (work, numSamples);
+        dsp[(size_t) c].processBlock (work, numSamples);
 
         // Hard-gate sub-threshold DSP output to prevent inaudible WDF residuals leaking.
         // If every sample is below ~-140 dBFS the whole block is zeroed; early-exit avoids
